Make IO::Tee compatible with IO::Wrap (and maybe other IO::* classes).
IO::Tee uses print as a function, thus the file handle must be a GLOB
reference. (The error message is "Not a GLOB reference...".) IO::Wrap
is a class with a print method, that should be usable as well.

In fact, with that change any class with file handle compatible methods
could be used with IO::Tee.